adding bootstrap files, initial work on v5 interface work on v5 interface, wrapper styling complete, table styling complete, form styling is wip

+++ views/admin/content/index.tpl
@@ -1,63 +1,46 @@
-{include file="inc_header.tpl" page_title="Content Pages" menu="content" page_style="fullContent"}
-{assign var=subMenu value="Content Pages"}
-{head}
-<script src="/scripts/dataTables/jquery.dataTables.min.js"></script>
-<script src="/scripts/dataTables/plugins/paging-plugin.js"></script>
-<script type="text/javascript">
-	$(function(){ldelim}
-		$('.dataTable').dataTable({ldelim}
-			/* DON'T CHANGE */
-			"sDom": 'rt<"dataTable-footer"flpi<"clear">',
-			"sPaginationType": "scrolling",
-			"bLengthChange": true,
-			/* CAN CHANGE */
-			"bStateSave": true, //whether to save a cookie with the current table state
-			"iDisplayLength": 10, //how many items to display on each page
-			"aaSorting": [[ 0, "asc" ]] //which column to sort by (0-X)
-		{rdelim});
-	{rdelim});
-</script>
-{/head}
+{$menu = "content"}{$subMenu = "Pages"}
+{include file="inc_header.tpl" sPageTitle="Content Pages"}
 
-<section id="content" class="content">
-	<header>
-		<h2>Content Pages</h2>
-		<a href="/admin/content/add/" title="Add Page" class="button">Add Page &raquo;</a>
-	</header>
+	<h1>Content Pages <a class="btn btn-primary pull-right" href="/admin/content/add/" title="Create Page"><i class="icon-plus icon-white"></i> Create Page</a></h1>
+	{include file="inc_alerts.tpl"}
 
-	<table class="dataTable">
+	<table class="data-table table table-striped">
 		<thead>
 			<tr>
 				<th>Title</th>
 				<th>URL</th>
-				<th></th>
+				<th>&nbsp;</th>
 			</tr>
 		</thead>
 		<tbody>
 			{foreach from=$aPages item=aPage}
 				<tr>
 					<td>{$aPage.title|clean_html}</td>
+					<td><a href="http://{$domain}/{$aPage.tag}/" target="new">http://{$domain}/{$aPage.tag}/</td>
 					<td>
-						{if $aPage.module == 0}
-							<a href="http://{$domain}/{$aPage.tag}/" target="new">http://{$domain}/{$aPage.tag}/</a>
-						{/if}
-					</td>
-					<td class="center">
-						<a href="/admin/content/edit/{$aPage.id}/" title="Edit Page">
-							<img src="/images/admin/icons/pencil.png" alt="edit icon">
-						</a>
-						{if $aPage.perminate != 1}
-							<a href="/admin/content/delete/{$aPage.id}/"
-							 onclick="return confirm_('Are you sure you would like to delete this page?');" title="Delete Page">
-								<img src="/images/admin/icons/bin_closed.png" alt="delete icon">
-							</a>
-						{else}
-							<img src="/images/admin/spacerIcon.png" alt="spacer">
+						<a href="/admin/content/edit/{$aPage.id}/" title="Edit Page"><i class="icon-pencil"></i></a>
+						{if $aPage.permanent != 1}
+							<a href="/admin/content/delete/{$aPage.id}/" title="Delete Page" onclick="return confirm('Are you sure you would like to delete: {$aPage.title}?');"><i class="icon-trash"></i></a>
 						{/if}
 					</td>
 				</tr>
 			{/foreach}
 		</tbody>
 	</table>
-</section>
+
+{footer}
+<script>
+$('.data-table').dataTable({
+	/* DON'T CHANGE */
+	"sDom": '<"dataTable-header"rf>t<"dataTable-footer"lip<"clear">',
+	"sPaginationType": "full_numbers",
+	"bLengthChange": false,
+	/* CAN CHANGE */
+	"bStateSave": true,
+	"aaSorting": [[ 0, "asc" ]], //which column to sort by (0-X)
+	"iDisplayLength": 10 //how many items to display per page
+});
+$('.dataTable-header').prepend('{foreach from=$aAdminFullMenu item=aMenu key=k}{if $k == $menu}{if $aMenu.menu|@count gt 1}<ul class="nav nav-pills">{foreach from=$aMenu.menu item=aItem}<li{if $subMenu == $aItem.text} class="active"{/if}><a href="{$aItem.link}" title="{$aItem.text}">{$aItem.text}</a></li>{/foreach}</ul>{/if}{/if}{/foreach}');
+</script>
+{/footer}
 {include file="inc_footer.tpl"}
